scripts/requirements.txt
pandas==2.2.3
python-slugify==8.0.1
Pillow==11.1.0
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1 